Move DummyInterfaceProperty
